+++ app/MicropubClient.php
@@ -4,7 +4,7 @@ namespace App;
 
 use Illuminate\Database\Eloquent\Model;
 
-class Client extends Model
+class MicropubClient extends Model
 {
     /**
      * The table associated with the model.