+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Version 0.0.3 (2013-06-09)
+  - Better tag normalization code organisation
+  - Remove `jonnybarnes/unicode-tools` dependency and clean up relevant code
+
 ## Version 0.0.2 (2016-05-25)
   - Fix issue#1: tagged notes page needs the tag from the URL normalizing.